feat: Crud completo na pagina de admin, componente tags, navBar e input refatorados.

+++ src/hooks/useProdutosMutation.ts
@@ -1,6 +1,17 @@
-import { useMutation, useQueryClient } from "@tanstack/react-query";
+import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 import { http } from "../http";
 import { IProdutos } from "../types/IProdutos";
+import useProdutos from "./useProdutos";
+
+export const useGetProdutos = (id: string) => {
+  return useQuery({
+    queryKey: ["produto", id],
+    queryFn: () => {
+      return http.get(`produtos?id=${id}`);
+    },
+    enabled: !!id,
+  });
+};
 
 const useProdutosMutation = () => {
   const queryClient = useQueryClient();
@@ -10,25 +21,27 @@ const useProdutosMutation = () => {
       action,
       value,
     }: {
-      id: string;
-      action: string;
-      value?: IProdutos;
+      id?: string;
+      action: "delete" | "new" | "edit";
+      value: Partial<IProdutos>; //? Permite passar apenas os campos necessários
     }) => {
-      console.log(id);
       if (action === "delete") {
-        console.log("oie");
-        return http.delete(`produtos/${id}`);
+        const deletar = confirm(`deseja deletar o produto ${value?.nome}`);
+        if (deletar) return http.delete(`produtos/${id}`);
+        return;
       }
       if (action === "new") {
-        return http.post(`produtos/${id}`, { ...value });
+        alert("Produto Criado!");
+        return http.post("produtos", { ...value });
       }
       if (action === "edit") {
+        alert("Produto Editado!");
         return http.put(`produtos/${id}`, { ...value });
       }
     },
 
     onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ["produtos"] });
+      queryClient.invalidateQueries({ queryKey: ["produtos-paginados"] });
     },
   });
 };
